Update CartView, Search Bar, BottomNavigationBar

+++ lib/model/cart.dart
@@ -5,15 +5,15 @@ class CartProvider extends ChangeNotifier {
   final List<Food> foods = [];
   void addToCart(Food food) {
     foods.add(food);
-     notifyListeners(); 
+    notifyListeners();
   }
 
   void removeFromCart(Food food) {
     foods.remove(food);
-     notifyListeners(); 
+    notifyListeners();
   }
 
-  double getTotal() {
+  int getTotal() {
     return foods.fold(0, (sum, food) => sum + food.price);
   }
 }
